xen-hvm: move common functions to hw/xen/xen-hvm-common.c
Extract common functionalities from xen-hvm.c and move them to
hw/xen/xen-hvm-common.c. These common functions are useful for creating
an IOREQ server.

Moved the common usable IOREQ creation part to a new function
xen_register_ioreq() which can be used by both x86 and ARM machines.

NOTE: This patch will break the build as the patch only involves moving
of functions. Build fixes will be in the next patch.

[Rebased to v7.0.0]
